Move code into `cell.module.view`
