+++ code/cell.module.view/src/Module/Module.create.ts
