+++ code/cell.module.view/src/Module/Module.events.ts
