+++ code/cell.module.view/src/Module/Module.fire.ts
