+++ code/cell.module.view/src/Module/Module.ts
