+++ code/cell.module.view/src/Module/types.ts
