+++ code/cell.module.view/src/common/libs.ts
@@ -1,2 +1,8 @@
 import { rx } from '@platform/util.value';
 export { rx };
+
+import { css, CssValue } from '@platform/css';
+export { css, CssValue };
+
+import { dispose } from '@platform/util.value';
+export { dispose };
