+++ code/cell.module.view/src/components/ModuleView.Frame/Frame.tsx
@@ -1,7 +1,8 @@
 import * as React from 'react';
 import { Subject } from 'rxjs';
 import { takeUntil, filter } from 'rxjs/operators';
-import { rx, css, CssValue, t, Module } from '../common';
+import { rx, css, CssValue, t } from '../../common';
+import { Module } from '../../Module';
 
 export type IModuleViewFrameProps = {
   bus: t.EventBus<any>;
