+++ code/cell.module.view/src/components/ModuleView.Frame/index.ts
