+++ code/cell.module.view/src/components/ModuleView/ModuleView.tsx
@@ -1,0 +1,9 @@
+import { ModuleViewFrame } from '../ModuleView.Frame';
+import { ModuleViewTree } from '../ModuleView.Tree';
+import { Module } from '../../Module';
+
+export const ModuleView = {
+  Module: Module,
+  Frame: ModuleViewFrame,
+  Tree: ModuleViewTree,
+};
