+++ code/cell.ui/package.json
@@ -14,18 +14,16 @@
   "dependencies": {
     "@platform/cell.client": "0.2.139",
     "@platform/cell.module": "0.0.6",
+    "@platform/cell.module.view": "0.0.3",
     "@platform/cell.types": "0.4.143",
     "@platform/libs": "0.2.2",
     "@platform/polyfill": "0.0.1",
     "@platform/react": "0.4.43",
-    "@platform/state": "0.7.5",
     "@platform/types": "0.5.3",
     "@platform/ui.button": "0.3.78",
     "@platform/ui.icon": "0.0.162",
     "@platform/ui.text": "0.3.74",
-    "@platform/ui.tree": "0.8.1",
     "@platform/util.string": "0.5.10",
-    "@platform/util.value": "0.7.30",
     "react-icons": "3.10.0"
   },
   "devDependencies": {
