--- code/cell.ui/src/components/ModuleView/ModuleView.tsx
@@ -1,9 +1,0 @@
-import { ModuleViewFrame } from './ModuleView.Frame';
-import { ModuleViewTree } from './ModuleView.Tree';
-import { Module } from '../../Module.view';
-
-export const ModuleView = {
-  Module: Module,
-  Frame: ModuleViewFrame,
-  Tree: ModuleViewTree,
-};
