--- code/cell.ui/src/components/ModuleView/common.ts
@@ -1,1 +1,0 @@
-export * from '../../common';
